+++ src/_prob.jl
@@ -3,8 +3,44 @@ import DD.BDD
 export prob
 export cprob
 
+"""
+    prob(ft::FTree, top::AbstractFTObject, env::Dict{Symbol,Tv}; type = :F, bddcache = Dict())::Tv where Tv <: Number
+    prob(ft::FTree, top::BDD.AbstractNode, env::Dict{Symbol,Tv}; type = :F, bddcache = Dict())::Tv where Tv <: Number
+
+Compute the probability of topevent
+
+## Arguments
+- `ft`: Context of fault tree
+- `top`: Top event node (AbstractFTObject or BDD node)
+- `env`: The dictionary to provide the probability of events
+- `type`: The symbol to determine whether the top event indicates failure or working. If type is :F, the top event is
+   the failure. Otherwise, if type is :G, the top event is working.
+- `bddcache`: The dictionary to store the intermediate computation results
+"""
+function prob(ft::FTree, top::AbstractFTObject, env::Dict{Symbol,Tv}; type = :F, bddcache = Dict())::Tv where Tv <: Number
+    prob(ft, ftbdd!(ft, top), env, type = type, bddcache = bddcache)
+end
+
 function prob(ft::FTree, f::BDD.AbstractNode, env::Dict{Symbol,Tv}; type = :F, bddcache = Dict())::Tv where Tv <: Number
     _prob(Val(:CDF), Val(type), ft, f, env, bddcache)
+end
+
+"""
+    cprob(ft::FTree, top::AbstractFTObject, env::Dict{Symbol,Tv}; type = :F, bddcache = Dict())::Tv where Tv <: Number
+    cprob(ft::FTree, top::BDD.AbstractNode, env::Dict{Symbol,Tv}; type = :F, bddcache = Dict())::Tv where Tv <: Number
+
+Compute the complementary probability of topevent
+
+## Arguments
+- `ft`: Context of fault tree
+- `top`: Top event node (AbstractFTObject or BDD node)
+- `env`: The dictionary to provide the probability of events
+- `type`: The symbol to determine whether the top event indicates failure or working. If type is :F, the top event is
+   the failure. Otherwise, if type is :G, the top event is working.
+- `bddcache`: The dictionary to store the intermediate computation results
+"""
+function cprob(ft::FTree, top::AbstractFTObject, env::Dict{Symbol,Tv}; type = :F, bddcache = Dict())::Tv where Tv <: Number
+    cprob(ft, ftbdd!(ft, top), env, type = type, bddcache = bddcache)
 end
 
 function cprob(ft::FTree, f::BDD.AbstractNode, env::Dict{Symbol,Tv}; type = :F, bddcache = Dict())::Tv where Tv <: Number