+++ test/_test_ftbdd.jl
@@ -1,16 +1,16 @@
 using DD.BDD
 
 @testset "FaultTreeBDD1" begin
-    top = ftbasic(:x) & ftbasic(:x)
     ft = FTree()
-    f = ftree(ft, top)
+    top = ftbasic(:x) & ftbasic(:x)
+    f = ftbdd!(ft, top)
     println(todot(f))
 end
 
 @testset "FaultTreeBDD2" begin
+    ft = FTree()
     x = ftbasic(:x)
     top = ftkofn(2, x, x, x)
-    ft = FTree()
-    f = ftree(ft, top)
+    f = ftbdd!(ft, top)
     println(todot(f))
 end